End of Section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -16,7 +17,7 @@ export class MessagesComponent implements OnInit {
   pageNumber: number = 1;  
   loading: boolean = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -39,10 +40,17 @@ export class MessagesComponent implements OnInit {
 
   deleteMessage(id: number)
   {
-    this.messageService.deleteMessage(id)
-      .subscribe(()=>{
-        //this.messages = this.messages.filter(item=>item.id != id);
-        this.messages.splice(this.messages.findIndex(m=>m.id == id),1);
+    this.confirmService.confirm("Confirmd delete message", "This cannot be undone")
+      .subscribe(result=>{
+        if(result)
+        {
+          //OK to delete
+          this.messageService.deleteMessage(id)
+          .subscribe(()=>{
+            //this.messages = this.messages.filter(item=>item.id != id);
+            this.messages.splice(this.messages.findIndex(m=>m.id == id),1);
+          });          
+        }
       });
   }
